% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyDescriptives.R
\name{ggSvyBoxplot}
\alias{ggSvyBoxplot}
\title{Survey Box plot}
\usage{
ggSvyBoxplot(design, variable, split = NULL, all.outliers = TRUE)
}
\arguments{
\item{design}{the survey design}

\item{variable}{the variable to plot}

\item{split}{NULL or one or more variables to split the boxplot by}

\item{all.outliers}{logical, whether to show all outliers or just the extreme ones, for compatability with \code{survey::svyboxplot}.}
}
\description{
Survey Box plot
}
\examples{
library(survey)
data(api)
apistrat$interaction <- factor(paste(apistrat$stype, apistrat$comp.imp, sep=" & "))
dstrat <- svydesign(id = ~1, strata = ~stype, weights = ~pw, data = apistrat, fpc = ~fpc)

svyboxplot(enroll~1,dstrat, plot = FALSE, all.outliers = TRUE, ylims = c(0, 3500))
ggSvyBoxplot("enroll", dstrat) + ggplot2::ylim(c(0, 3500))

svyboxplot(enroll~stype,dstrat,all.outliers=TRUE)
ggSvyBoxplot("enroll", dstrat, "stype")

svyboxplot(enroll~stype * comp.imp,dstrat,all.outliers = TRUE) # fails
svyboxplot(enroll~interaction,dstrat,all.outliers = TRUE) # workaround
ggSvyBoxplot("enroll", dstrat, c("stype", "comp.imp"))

}
